33
- removed old code for M commands from ProcessLine()
- added line numbers to gcode file transmits
- added CRC to lines sent from file to robot via GUI
- added M110 (set line number)
- added line number check in firmware
- added CRC check in firmware

+++ java/src/Makelangelo/GCodeFile.java
@@ -15,8 +15,8 @@ import java.util.Scanner;
  *
  */
 public class GCodeFile {
-    public long linesTotal=0;
-	public long linesProcessed=0;
+    public int linesTotal=0;
+	public int linesProcessed=0;
 	public boolean fileOpened=false;
 	public ArrayList<String> lines;
 	public float estimated_time=0;
